backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2

mysqlclient==2.2.4

django-environ==0.11.2
django-cors-headers==4.6.0

djangorestframework-simplejwt==5.3.1
